make the delete row will check the broken image is null image or not

+++ project/customer_delete.php
@@ -1,11 +1,11 @@
 <?php
 // include database connection
 include 'config/database.php';
-try {     
+try {
     // get record ID
     // isset() is a PHP function used to verify if a value is there or not
-    $id=isset($_GET['id']) ? $_GET['id'] :  die('ERROR: Record ID not found.');
-    $select ="SELECT Username AS customername, image FROM customers where id=:id";
+    $id = isset($_GET['id']) ? $_GET['id'] :  die('ERROR: Record ID not found.');
+    $select = "SELECT Username AS customername, image FROM customers where id=:id";
     $stmt = $con->prepare($select);
     $stmt->bindParam(":id", $id);
     $stmt->execute();
@@ -13,31 +13,32 @@ try {
     extract($row);
 
     // delete query
-    $check ="SELECT customer_order FROM order_summary WHERE customer_order=:customer_order";
+    $check = "SELECT customer_order FROM order_summary WHERE customer_order=:customer_order";
     $stmt = $con->prepare($check);
     $stmt->bindParam(":customer_order", $customername);
     $stmt->execute();
     $count = $stmt->rowCount();
-    
-    if($count > 0){
+
+    if ($count > 0) {
         header("Location: customer_read.php?action=nodelete");
-    }else{
+    } else {
         $query = "DELETE FROM customers WHERE id = ?";
         $stmt = $con->prepare($query);
         $stmt->bindParam(1, $id);
-         
-        if($stmt->execute()){
-            unlink("uploads/customer/" . $row['image']);
-            // redirect to read records page and
-            // tell the user record was deleted
-            header('Location: customer_read.php?action=deleted');
-        }else{
+
+        if ($stmt->execute()) {
+            if ($row['image'] != "customer/profile.jpg") {
+                unlink("uploads/customer/" . $row['image']);
+                // redirect to read records page and
+                // tell the user record was deleted
+                header('Location: customer_read.php?action=deleted');
+            }
+        } else {
             die('Unable to delete record.');
         }
     }
-    
 }
 // show error
-catch(PDOException $exception){
+catch (PDOException $exception) {
     die('ERROR: ' . $exception->getMessage());
 }
